+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/MainConfigScreen.java
@@ -2,7 +2,6 @@ package club.iananderson.seasonhud.client.gui.screens;
 
 import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.client.gui.Location;
-import club.iananderson.seasonhud.client.gui.ShowDay;
 import club.iananderson.seasonhud.client.gui.components.buttons.MenuButton;
 import club.iananderson.seasonhud.client.gui.components.buttons.MenuButton.MenuButtons;
 import club.iananderson.seasonhud.client.gui.components.sliders.BasicSlider;
@@ -23,44 +22,62 @@ import net.minecraft.network.chat.MutableComponent;
 import org.jetbrains.annotations.NotNull;
 
 public class MainConfigScreen extends SeasonHudScreen {
-  private static final Component SCREEN_TITLE = Component.translatable("menu.seasonhud.title");
-  private static final Component JOURNEYMAP = Component.translatable("menu.seasonhud.title.journeymap");
+  private static final Component SCREEN_TITLE = Component.translatable("menu.seasonhud.options.title");
+  private static final Component JOURNEYMAP = Component.translatable("menu.seasonhud.options.journeymap.title");
   private final List<AbstractWidget> optionButtons = new ArrayList<>();
+  private boolean enableMod;
+  private Location hudLocation;
+  private int xSliderInt;
+  private int ySliderInt;
+  private boolean showMinimapHidden;
+  private boolean enableMinimapIntegration;
+  private boolean needCalendar;
+  private boolean journeyMapAboveMap;
+  private boolean journeyMapMacOS;
   private CycleButton<Boolean> enableModButton;
   private CycleButton<Location> hudLocationButton;
-  private HudOffsetSlider xOffsetSlider;
-  private HudOffsetSlider yOffsetSlider;
-  private CycleButton<ShowDay> showDayButton;
-  private CycleButton<Boolean> showSubSeasonButton;
-  private CycleButton<Boolean> showTropicalSeasonButton;
-  private CycleButton<Boolean> needCalendarButton;
+  private HudOffsetSlider xSlider;
+  private HudOffsetSlider ySlider;
   private CycleButton<Boolean> showMinimapHiddenButton;
   private CycleButton<Boolean> enableMinimapIntegrationButton;
+  private CycleButton<Boolean> needCalendarButton;
   private CycleButton<Boolean> journeyMapAboveMapButton;
   private CycleButton<Boolean> journeyMapMacOSButton;
 
+
   public MainConfigScreen() {
     super(null, SCREEN_TITLE);
+    loadConfig();
   }
 
   public static MainConfigScreen getInstance() {
     return new MainConfigScreen();
   }
 
-  public void saveConfig() {
-    Config.setEnableMod(enableModButton.getValue());
-    Config.setHudLocation(hudLocationButton.getValue());
-    Config.setHudX(xOffsetSlider.getValueInt());
-    Config.setHudY(yOffsetSlider.getValueInt());
-    Config.setShowDay(showDayButton.getValue());
-    Config.setShowSubSeason(showSubSeasonButton.getValue());
-    Config.setShowTropicalSeason(showTropicalSeasonButton.getValue());
-    Config.setNeedCalendar(needCalendarButton.getValue());
-    Config.setShowDefaultWhenMinimapHidden(showMinimapHiddenButton.getValue());
-    Config.setEnableMinimapIntegration(enableMinimapIntegrationButton.getValue());
+  public void loadConfig() {
+    enableMod = Config.getEnableMod();
+    hudLocation = Config.getHudLocation();
+    xSliderInt = Config.getHudX();
+    ySliderInt = Config.getHudY();
+    showMinimapHidden = Config.getShowDefaultWhenMinimapHidden();
+    enableMinimapIntegration = Config.getEnableMinimapIntegration();
+    needCalendar = Config.getNeedCalendar();
     if (CurrentMinimap.journeyMapLoaded()) {
-      Config.setEnableMod(journeyMapAboveMapButton.getValue());
-      Config.setEnableMod(journeyMapMacOSButton.getValue());
+      journeyMapAboveMap = Config.getEnableMod();
+      journeyMapMacOS = Config.getEnableMod();
+    }
+  }
+
+  public void saveConfig() {
+    Config.setEnableMod(enableMod);
+    Config.setHudLocation(hudLocation);
+    Config.setHudX(xSliderInt);
+    Config.setHudY(ySliderInt);
+    Config.setShowDefaultWhenMinimapHidden(showMinimapHidden);
+    Config.setNeedCalendar(needCalendar);
+    if (CurrentMinimap.journeyMapLoaded()) {
+      Config.setEnableMod(journeyMapAboveMap);
+      Config.setEnableMod(journeyMapMacOS);
     }
   }
 
@@ -68,6 +85,12 @@ public class MainConfigScreen extends SeasonHudScreen {
   public void onDone() {
     saveConfig();
     super.onDone();
+  }
+
+  @Override
+  public void onClose() {
+    Config.setEnableMinimapIntegration(enableMinimapIntegration);
+    super.onClose();
   }
 
   @Override
@@ -81,117 +104,98 @@ public class MainConfigScreen extends SeasonHudScreen {
     }
 
     hudLocationButton.active = Common.drawDefaultHud();
-    xOffsetSlider.active = hudLocationButton.getValue() == Location.TOP_LEFT && Common.drawDefaultHud();
-    yOffsetSlider.active = hudLocationButton.getValue() == Location.TOP_LEFT && Common.drawDefaultHud();
+    xSlider.active = hudLocationButton.getValue() == Location.TOP_LEFT && Common.drawDefaultHud();
+    ySlider.active = hudLocationButton.getValue() == Location.TOP_LEFT && Common.drawDefaultHud();
   }
 
   @Override
   public void init() {
-    MutableComponent seasonCombined = CurrentSeason.getInstance(this.minecraft).getSeasonHudText();
+    super.init();
 
+    MutableComponent seasonCombined = CurrentSeason.getInstance(this.minecraft).getSeasonHudText();
     double componentWidth = this.font.width(seasonCombined);
     double componentHeight = this.font.lineHeight;
 
-    super.init();
-
-    int buttonStartY = MENU_PADDING;
-    int yOffset = BUTTON_HEIGHT + BUTTON_PADDING;
-
     int row = 0;
-    //noinspection ConstantValue
-    enableModButton = CycleButton.onOffBuilder(Config.getEnableMod())
+    enableModButton = CycleButton.onOffBuilder(enableMod)
         .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.enableMod")))
         .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.enableMod"));
+                Component.translatable("menu.seasonhud.button.enableMod"), (b, val) -> enableMod = val);
 
-    MenuButton colorButton = MenuButton.builder(MenuButtons.COLORS, press -> ColorScreen.open(this))
-        .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.tooltip.color")))
-        .withPos(rightButtonX, (buttonStartY + (row * yOffset)))
-        .withWidth(BUTTON_WIDTH)
-        .build();
+    enableMinimapIntegrationButton = CycleButton.onOffBuilder(enableMinimapIntegration)
+        .withTooltip(
+            object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.enableMinimapIntegration")))
+        .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
+                Component.translatable("menu.seasonhud.button.enableMinimapIntegration"),
+                (b, val) -> Config.setEnableMinimapIntegration(val));
 
     row = 1;
     hudLocationButton = CycleButton.builder(Location::getLocationName)
         .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.hudLocation")))
         .withValues(Location.TOP_LEFT, Location.TOP_CENTER, Location.TOP_RIGHT, Location.BOTTOM_LEFT,
-                    Location.BOTTOM_RIGHT)
-        .withInitialValue(Config.getHudLocation())
+                    Location.BOTTOM_RIGHT).withInitialValue(hudLocation)
         .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.hudLocation"));
+                Component.translatable("menu.seasonhud.button.hudLocation"), (b, val) -> hudLocation = val);
 
-    xOffsetSlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud.slider.xOffset"))
+    xSlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud.slider.xOffset"))
         .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.tooltip.xOffset")))
-        .withValueRange(0, this.width - componentWidth)
-        .withInitialValue(Config.getHudX())
+        .withValueRange(0, this.width - componentWidth).withInitialValue(xSliderInt)
         .withDefaultValue(Config.DEFAULT_X_OFFSET)
         .withBounds(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH / 2 - BasicSlider.SLIDER_PADDING,
                     BUTTON_HEIGHT)
         .build();
 
-    yOffsetSlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud" + ".slider.yOffset"))
+    ySlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud" + ".slider.yOffset"))
         .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.tooltip.yOffset")))
-        .withValueRange(0, this.height - componentHeight)
-        .withInitialValue(Config.getHudY())
+        .withValueRange(0, this.height - componentHeight).withInitialValue(ySliderInt)
         .withDefaultValue(Config.DEFAULT_Y_OFFSET)
         .withBounds(rightButtonX + BUTTON_WIDTH / 2 + BasicSlider.SLIDER_PADDING, (buttonStartY + (row * yOffset)),
                     BUTTON_WIDTH / 2 - BasicSlider.SLIDER_PADDING, BUTTON_HEIGHT)
         .build();
 
     row = 2;
-    showDayButton = CycleButton.builder(ShowDay::getDayDisplayName)
-        .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.showDay")))
-        .withValues(ShowDay.getValues())
-        .withInitialValue(Config.getShowDay())
-        .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.showDay"));
+    MenuButton seasonButton = MenuButton.builder(MenuButtons.SEASON, b -> SeasonOptionsScreen.getInstance(this).open())
+        .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.tooltip.color")))
+        .withPos(leftButtonX, (buttonStartY + (row * yOffset)))
+        .withWidth(BUTTON_WIDTH)
+        .build();
 
-    showSubSeasonButton = CycleButton.onOffBuilder(Config.getShowSubSeason())
-        .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.showSubSeason")))
-        .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.showSubSeason"));
+    MenuButton colorButton = MenuButton.builder(MenuButtons.COLORS, b -> ColorScreen.getInstance(this).open())
+        .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.tooltip.color")))
+        .withPos(rightButtonX, (buttonStartY + (row * yOffset)))
+        .withWidth(BUTTON_WIDTH)
+        .build();
 
     row = 3;
-    showTropicalSeasonButton = CycleButton.onOffBuilder(Config.getShowTropicalSeason())
-        .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.showTropicalSeason")))
-        .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.showTropicalSeason"));
-
-    needCalendarButton = CycleButton.onOffBuilder(Config.getNeedCalendar())
-        .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.needCalendar")))
-        .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.needCalendar"));
-
-    row = 4;
-    showMinimapHiddenButton = CycleButton.onOffBuilder(Config.getShowDefaultWhenMinimapHidden())
+    showMinimapHiddenButton = CycleButton.onOffBuilder(showMinimapHidden)
         .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.showMinimapHidden")))
         .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.showMinimapHidden"));
+                Component.translatable("menu.seasonhud.button.showMinimapHidden"), (b, val) -> showMinimapHidden = val);
 
-    enableMinimapIntegrationButton = CycleButton.onOffBuilder(Config.getEnableMinimapIntegration())
-        .withTooltip(
-            object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.enableMinimapIntegration")))
+    needCalendarButton = CycleButton.onOffBuilder(needCalendar)
+        .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.needCalendar")))
         .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.button.enableMinimapIntegration"));
+                Component.translatable("menu.seasonhud.button.needCalendar"), (b, val) -> needCalendar = val);
+
+    widgets.addAll(Arrays.asList(enableModButton, enableMinimapIntegrationButton, hudLocationButton, xSlider, ySlider,
+                                 seasonButton, colorButton, showMinimapHiddenButton, needCalendarButton));
 
     if (Services.PLATFORM.isModLoaded("journeymap")) {
       row += 2; //6
-      journeyMapAboveMapButton = CycleButton.onOffBuilder(Config.getJourneyMapAboveMap())
+      journeyMapAboveMapButton = CycleButton.onOffBuilder(journeyMapAboveMap)
           .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.journeyMapAboveMap")))
           .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.button.journeyMapAboveMap"));
+                  Component.translatable("menu.seasonhud.button.journeyMapAboveMap"),
+                  (b, val) -> journeyMapAboveMap = val);
 
-      journeyMapMacOSButton = CycleButton.onOffBuilder(Config.getJourneyMapMacOS())
+      journeyMapMacOSButton = CycleButton.onOffBuilder(journeyMapMacOS)
           .withTooltip(object -> Tooltip.create(Component.translatable("menu.seasonhud.tooltip.journeyMapMacOS")))
           .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.button.journeyMapMacOS"));
+                  Component.translatable("menu.seasonhud.button.journeyMapMacOS"), (b, val) -> journeyMapMacOS = val);
 
       widgets.add(journeyMapAboveMapButton);
       widgets.add(journeyMapMacOSButton);
     }
-    widgets.addAll(
-        Arrays.asList(enableModButton, colorButton, hudLocationButton, xOffsetSlider, yOffsetSlider, showDayButton,
-                      showSubSeasonButton, showTropicalSeasonButton, needCalendarButton, showMinimapHiddenButton,
-                      enableMinimapIntegrationButton));
 
     widgets.forEach(this::addRenderableWidget);
   }